dagknight: restore pre-PR placement of DagknightData and GroupMetadata
